Creando el servicio REST con el método HTTP GET y POST para OfficeAssignments

+++ University/University.BL/DTOs/DepartmentDTO.cs
@@ -8,7 +8,7 @@ namespace University.BL.DTOs
     {
         public int DepartmentID { get; set; }
         public string Name { get; set; }
-        public float Budget { get; set; }
+        public Decimal Budget { get; set; }
 
         [Display(Name = "Start Date")]
         public DateTime StartDate { get; set; }
